+++ zephyrapp/game/src/cowtools/cat_math.c
@@ -20,7 +20,7 @@ float inv_lerp(float t, float a, float b)
 
 int quantize(float t, float range, int steps)
 {
-	return clamp(round((t / range) * (float) (steps - 1)), 0, steps - 1);
+	return clamp(roundf((t / range) * (float) (steps - 1)), 0, steps - 1);
 }
 
 
@@ -85,7 +85,7 @@ float CAT_vec2_mag2(CAT_vec2 a)
 
 CAT_vec2 CAT_vec2_unit(CAT_vec2 a)
 {
-	float inv_mag = 1.0f/sqrt(a.x*a.x + a.y*a.y);
+	float inv_mag = 1.0f/sqrtf(a.x*a.x + a.y*a.y);
 	return (CAT_vec2) {a.x*inv_mag, a.y*inv_mag};
 }
 
@@ -228,7 +228,7 @@ CAT_vec4 CAT_vec4_sub(CAT_vec4 u, CAT_vec4 v)
 
 CAT_vec4 CAT_vec4_normalize(CAT_vec4 v)
 {
-	float len = sqrt(v.x * v.x + v.y * v.y + v.z * v.z + v.w * v.w);
+	float len = sqrtf(v.x * v.x + v.y * v.y + v.z * v.z + v.w * v.w);
 	float s = 1.0f / len;
 	return (CAT_vec4) {v.x * s, v.y * s, v.z * s, v.w * s};
 }
